Move editTodo function to module
