+++ src/deleteTodo.js
@@ -4,8 +4,8 @@ import addToStorage from './addToStorage.js';
 const deleteCompletedTodoItem = (tasks) => {
   const clearAll = document.getElementsByClassName('clearAll');
   clearAll[0].addEventListener('click', () => {
-    function finish(el) {
-      return !el.completed === true;
+    function finish(todo) {
+      return !todo.completed === true;
     }
     tasks = tasks.filter(finish);
     for (let i = 0; i < tasks.length; i += 1) {
@@ -20,9 +20,8 @@ const deleteOnlyOne = (tasks) => {
   const trashIcon = document.querySelectorAll('.fa-trash-alt');
 
   trashIcon.forEach((icon) => icon.addEventListener('click', () => {
-    const { id } = icon.parentNode.id;
-
-    tasks.splice(id, 1);
+    const todoItem = icon.parentNode;
+    tasks.splice(todoItem.id, 1);
     for (let i = 0; i < tasks.length; i += 1) {
       tasks[i].index = i;
     }
