+++ src/style.css
@@ -163,7 +163,7 @@ input {
 }
 
 input[type='checkbox'] {
-  transform: scale(1.5);
+  transform: scale(1.6);
 }
 
 .item {
